+++ src/app/faq/page.tsx
@@ -1,4 +1,4 @@
-import StandardPageLayout from "@/components/common/StandardPageLayout";
+import StandardPageLayout from "@/components/StandardPageLayout";
 
 export default function FaqPage() {
     return <StandardPageLayout>FaqPage</StandardPageLayout>;
