+++ src/app/projects/page.tsx
@@ -1,4 +1,4 @@
-import StandardPageLayout from "@/components/common/StandardPageLayout";
+import StandardPageLayout from "@/components/StandardPageLayout";
 
 export default function ProjectsPage() {
     return <StandardPageLayout>ProjectsPage</StandardPageLayout>;
